integration: Refactor restore related tests
1. Rename the test file to `test_restore.py`
2. Remove restore inc command and the related functions
3. Separate the failure test cases from test_restore_incrementally().
Since sync agent server will record and check the previous error
before starting a new backup restore, and typically the replica
failed backup restore should beomce unavailable. We cannot blindly
reuse the failed replicas in test_restore_incrementally()
4. Add cmd 'verify-rebuild-replica'
5. Add expansion and rebuilding test for DR volume.
6. Add a simple restore & rebuild test.

Longhorn #1279
